+++ configs/database.js
@@ -1,0 +1,8 @@
+/*!
+ * Copyright (c) 2023-2024 Digital Bazaar, Inc. All rights reserved.
+ */
+import {config} from '@bedrock/core';
+import '@bedrock/mongodb';
+
+// mongodb config
+config.mongodb.name = 'opencred_localhost';